Add module to read from MIDI file
This code reads a MIDI file and gets the note data in the form of a list
of dictionaries containing value of note and note duration. Least count
of note duration is 0.5 (For fast notes).

Has logic which will eliminate noise from MIDI file, gets notes
according to how portamento (simultaneous notes) notes are stored.